requirements.txt
fastapi
uvicorn
openai
python-docx